move from wchar_t to char32_t - because wchar_t isn't quite platform independantly the same size of 4 bytes ;-(
Also eliminating some compile warnings along the way.

+++ src/terminal/OutputGenerator.cpp
@@ -117,7 +117,7 @@ void OutputGenerator::sgr_add(int n)
 
 void OutputGenerator::operator()(Command const& command)
 {
-    auto const pairOrNone = [](int _default, int _a, int _b) -> string {
+    auto const pairOrNone = [](size_t _default, size_t _a, size_t _b) -> string {
         if (_a == _default && _b == _default)
             return "";
         else if (_a == _default)
